VAMPr dataset heatmap
This heatmap is based on the VAMPr dataset and created from the data in the `antibiogram_phenotype.txt` file. The heatmap shows the associations of the microorganisms used to the antibiotics evaluated in this study and indicates whether the organism is resistant/susceptible to a particular antibiotic. 

From this heatmap, it can be seen that **Escherichia coli** has more datapoints than other organisms and an even distribution of the phenotypes, making it a more suitable candidate to be used for our project.